Pass ijson_parse the length of the output buffer.
Passing the ijson_parse the length of the output buffer and the length
of the internal stack allows ijson_parse to check that all of its memory
accesses are valid. It also allows devices that have constraints on
available memory to parse JSON where number of nodes is small without
having to overallocate a large buffer.